refactor(core): move HTML-in-Canvas interaction managers into core
Relocate HtmlInteractionManager and HtmlRaycastInteractionManager from addons into a single core feature folder (Materials/Textures/HTML/) alongside HtmlTexture, mirroring core/Behaviors: opinionated, picking-dependent interaction helpers that are side-effect-free and tree-shaken unless imported. The three-html-render polyfill installer stays in addons as the external-package recipe. Tests moved into core; no manifest drift; side-effects sync clean.
