+++ packages/dev/addons/src/htmlInCanvas/index.ts
@@ -1,5 +1,3 @@
-import { HtmlInteractionManager, ComputeOverlayCssTransform, type IHtmlInteractionManagerOptions } from "./htmlInteractionManager";
-import { HtmlRaycastInteractionManager, GetElementPixelFromUv, type IHtmlRaycastInteractionManagerOptions } from "./htmlRaycastInteractionManager";
 import {
     InstallHtmlInCanvasPolyfill,
     UninstallHtmlInCanvasPolyfill,
@@ -9,6 +7,5 @@ import {
 } from "./polyfill";
 
 // Export public classes and functions
-export { HtmlInteractionManager, ComputeOverlayCssTransform, HtmlRaycastInteractionManager, GetElementPixelFromUv };
 export { InstallHtmlInCanvasPolyfill, UninstallHtmlInCanvasPolyfill, IsHtmlInCanvasSupportedNatively };
-export type { IHtmlInteractionManagerOptions, IHtmlRaycastInteractionManagerOptions, IInstallHtmlInCanvasPolyfillOptions, IHtmlInCanvasPolyfillModule };
+export type { IInstallHtmlInCanvasPolyfillOptions, IHtmlInCanvasPolyfillModule };
