+++ packages/dev/core/src/Materials/Textures/HTML/index.ts
@@ -1,0 +1,3 @@
+export * from "./htmlTexture";
+export * from "./htmlInteractionManager";
+export * from "./htmlRaycastInteractionManager";
