+++ packages/dev/core/src/Materials/Textures/HTML/pure.ts
@@ -1,0 +1,4 @@
+/** Pure barrel — re-exports only side-effect-free modules */
+export * from "./htmlTexture";
+export * from "./htmlInteractionManager";
+export * from "./htmlRaycastInteractionManager";
