+++ packages/dev/core/src/Materials/Textures/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-restricted-imports */
 export * from "./baseTexture";
-// eslint-disable-next-line import/export
+
 export * from "./baseTexture.polynomial";
 export * from "./colorGradingTexture";
 export * from "./cubeTexture";
@@ -12,7 +12,7 @@ export * from "./envCubeTexture";
 export * from "./hdrCubeTexture";
 export * from "./exrCubeTexture";
 export * from "./htmlElementTexture";
-export * from "./htmlTexture";
+export * from "./HTML/index";
 export * from "./internalTexture";
 export * from "./Loaders/index";
 export * from "./mirrorTexture";
